requirements.txt
flask
pdfminer.six
cohere
python-docx
youtube_transcript_api
